+++ test/array-spec.js
@@ -1,4 +1,5 @@
 var testHelper = require('../etc/test-helper.js');
+var errorFormatter = testHelper.validationErrorFormatter;
 
 describe('Array validation type', function() {
   beforeEach(function() {
@@ -21,7 +22,7 @@ describe('Array validation type', function() {
         lengthValidationProp: [ 'foo' ]
       };
 
-      testHelper.verifyDocumentNotCreated(doc, 'arrayDoc', 'length of item "lengthValidationProp" must not be less than 2');
+      testHelper.verifyDocumentNotCreated(doc, 'arrayDoc', errorFormatter.minimumLengthViolation('lengthValidationProp', 2));
     });
 
     it('cannot create a doc with an array that is longer than the maximum length', function() {
@@ -30,7 +31,7 @@ describe('Array validation type', function() {
         lengthValidationProp: [ 'foo', 'bar', 'baz' ]
       };
 
-      testHelper.verifyDocumentNotCreated(doc, 'arrayDoc', 'length of item "lengthValidationProp" must not be greater than 2');
+      testHelper.verifyDocumentNotCreated(doc, 'arrayDoc', errorFormatter.maximumLengthViolation('lengthValidationProp', 2));
     });
   });
 });