+++ test/resources/date-doc-definitions.js
@@ -9,6 +9,9 @@
         type: 'date',
         minimumValue: '2016-06-23',
         maximumValue: '2016-06-23'
+      },
+      formatValidationProp: {
+        type: 'date'
       }
     }
   }