+++ test/string-spec.js
@@ -1,4 +1,5 @@
 var testHelper = require('../etc/test-helper.js');
+var errorFormatter = testHelper.validationErrorFormatter;
 
 describe('String validation type', function() {
   beforeEach(function() {
@@ -21,7 +22,7 @@ describe('String validation type', function() {
         lengthValidationProp: 'fo'
       };
 
-      testHelper.verifyDocumentNotCreated(doc, 'stringDoc', 'length of item "lengthValidationProp" must not be less than 3');
+      testHelper.verifyDocumentNotCreated(doc, 'stringDoc', errorFormatter.minimumLengthViolation('lengthValidationProp', 3));
     });
 
     it('cannot create a doc with a string that is longer than the maximum length', function() {
@@ -30,7 +31,7 @@ describe('String validation type', function() {
         lengthValidationProp: 'foob'
       };
 
-      testHelper.verifyDocumentNotCreated(doc, 'stringDoc', 'length of item "lengthValidationProp" must not be greater than 3');
+      testHelper.verifyDocumentNotCreated(doc, 'stringDoc', errorFormatter.maximumLengthViolation('lengthValidationProp', 3));
     });
   });
 });